En comments.js se crean botones de eliminar y editar, se le da funcionalidad al boton de eliminar.
