+++ src/view/posts.js
@@ -1,4 +1,4 @@
-import { comments } from './comments.js';
+import { comments } from '../componentes/comments.js';
 import { header } from '../componentes/header.js';
 
 export const postsScreen = () => {
@@ -19,7 +19,7 @@ export const postsScreen = () => {
       <textarea id="postText" placeholder="Escribe un comentario"></textarea>
       <button id="btnPost" class="postButton">Comentar</button>
     </div>
-  </main>
+   </main>
 `;
   const divPosts = document.createElement('div');
   divPosts.setAttribute('class', 'viewPosts');
@@ -28,12 +28,10 @@ export const postsScreen = () => {
   divPostsContainer.appendChild(header());
   divPostsContainer.appendChild(divPosts);
   const btnPost = divPosts.querySelector('#btnPost');
-
   divPostsContainer.appendChild(comments());
 
   btnPost.addEventListener('click', () => {
     const postComment = divPosts.querySelector('#postText').value;
-    const comment = divPosts.querySelector('#postedComments');
 
     const db = firebase.firestore();
     db.collection('posts').add({
@@ -41,12 +39,11 @@ export const postsScreen = () => {
     })
       .then((docRef) => {
         console.log('Document written with ID: ', docRef.id);
-        comment.innerHTML = postComment;
+        divPosts.querySelector('#postText').value = '';
       })
       .catch((error) => {
         console.error('Error adding document: ', error);
       });
   });
-
   return divPostsContainer;
 };
